jobs/reportwriter/requirements.txt
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
